works out _get_class lookup construction
